[MWPW-132212] Template v2 Search Feature Update (#870)
* search function done for EN

* reinstate sheet powered page validation

* findMatchExistingSEOPage col name update

* redirectSearch col name update
